tuancv: update search anatytics
