+++ src/components/dropdown/index.js
@@ -1,3 +1,4 @@
 export { default as DropdownUsers } from './dropdownusers';
+export { default as SearchAnalytics } from './searchanalytics';
 export { default as DropdownLanguage } from './dropdownlanguage';
 export { default as DropdownNotifications } from './dropdownnotifications';
